+++ RELEASE_NOTES.md
@@ -1,8 +1,22 @@
-## [Rust SDK] 0.2.7 - 2025-07-24
+## [CLI] 0.0.32 - 2025-07-24
 
 ### Changes
 
+- release: Rust SDK v0.2.4 (#51)
+- release: Rust SDK v0.2.5 (#52)
+- release: Rust SDK v0.2.6 (#53)
+- release: Rust SDK v0.2.7 (#54)
+- release: TypeScript SDK v0.2.4 (#56)
+- release: mcp-gateway v0.0.5 (#57)
+- ✨ feat: Improve install.sh script. (#58)
+- 🐛 fix: exclude templates/examples from cargo update
+- 🐛 fix: install.sh interactive prompts
+- 🐛 fix: install.sh script
+- 🐛 fix: mcp-gateway rust sdk dep
+- 🐛 fix: rust sdk release
+- 🐛 fix: sdk release
 
 ### Contributors
 
+- Ian McDonald
 - bowlofarugula